toPng function
