+++ src/Parser.js
@@ -62,4 +62,31 @@ Parser.prototype.parse = function() {
   this.stream.input = null;
 };
 
+Parser.prototype.toPng = function(){
+  var height = this.header.rows;
+  var width = this.header.columns;
+
+  //create empty array to get pixel data
+  var canvas = document.createElement('canvas');
+  canvas.width = width;
+  canvas.height = height;
+  let context = canvas.getContext('2d');
+  let imageData = context.createImageData(width, height);
+  let pixelData = imageData.data;
+
+  //copy in colors
+  var colors = this.imageData.createColor(this.header, this.colorModeData);
+  for(let y = 0; y < height; ++y) {
+    for(let x = 0; x < width; ++x) {
+      let index = (y * width + x);
+      pixelData[index * 4    ] = colors[0][index];
+      pixelData[index * 4 + 1] = colors[1][index];
+      pixelData[index * 4 + 2] = colors[2][index];
+      pixelData[index * 4 + 3] = 255;
+    }
+  }
+  context.putImageData(imageData, 0, 0, 0, 0, width, height);
+  return canvas.toDataURL('image/png');
+}
+
 module.exports = Parser;
